fix: restore required narrowing casts for OpenSSL/BIO API contracts
Restore static_cast<int> for security-critical APIs:
- totp_secret_encryption.cpp: BIO_write, BIO_read, EVP_EncryptUpdate, EVP_DecryptUpdate
- saml_authenticator.cpp: BIO_write, BIO_read
- jwt_validator.cpp: BIO_read
- passkey_authenticator.cpp: BN_bin2bn (RSA key material)
- process_mining.cpp: n_variants = static_cast<int>(variant_keys.size())
- forecasting.cpp: total_ar = static_cast<int>(ar_lags.size())

+++ src/analytics/forecasting.cpp
@@ -910,7 +910,7 @@ SARIMAParams fitSARIMA(const std::vector<double> &y, int p, int d, int q, int P,
 
     // Build OLS design matrix for AR via Yule-Walker generalisation
     // (simple OLS regression of yc[t] on yc[t-lag] for lag in ar_lags)
-    int total_ar = ar_lags.size();
+    int total_ar = static_cast<int>(ar_lags.size());
     size_t n = yc.size();
     int max_lag  = ar_lags.empty() ? 0 : ar_lags.back();
     if (max_lag < 1 || static_cast<int>(n) <= max_lag + 1) {
